Product price value type refactor

+++ Models/Product.cs
@@ -3,7 +3,7 @@
     public class Product
     {
         public string Name { get; }
-        public float Price { get; }
+        public int Price { get; }
         public ProductType Type { get; }
         public DateTime SubscriptionExpirationDate { get; }
     }
